Allow to configure the channels on the job functions
With 2 new models: queue.job.channel and queue.job.function
The channels are a tree starting from 'root' which is a mandatory channel and
each job function (decorated with @job) can be assigned to a channel.

The channels names should correspond to the one used when the runner is launched.